requirements.txt
Pillow~=9.5
PyHyphen==4.0.3
pytz